moved text sharing broadcast receiver to it's own class
